Edited Script up to 'Talk to Maya'
The talk to maya section is completed, not including the final choices of that section before admission.
I've included a lot of sprite expressions, but I guess just include the ones we end up having. Music, I think we should figure out.

Saw some minor tense errors in the weapon choice for fist, and the second line of the  game. So edited those.